fix!: A bridge now must be configured with a unique ID
Before, the bridge ID was the bridge's MAC that was retrieved on successful connection. The problem with this approach is that if HB restarts and for some reason the configured bridge is not accessible, all the bridge's cached devices where removed. When the bridge becomes accessible again, they're created again, but the changes made to the accessories in HB are now lost.

Instead of using the bridge's MAC as an id, the id is now configured by the user. This means that regardless of the bridges successful/failed connection we know the id and we know if any of the cached accessories belong to any of the configured bridges. If we see an accessory that belong to one of the configured bridges we don't remove it. If the bridge is offline (not accessible) we mark the accessory with `StatusFault=1` to indicate it's not accessible.